[FLINK-8978] Stateful generic stream job upgrade e2e test
