+++ flink-end-to-end-tests/pom.xml
@@ -41,6 +41,7 @@ under the License.
 		<module>flink-stream-sql-test</module>
 		<module>flink-bucketing-sink-test</module>
 		<module>flink-distributed-cache-via-blob-test</module>
+		<module>flink-stream-stateful-job-upgrade-test</module>
 	</modules>
 
 </project>
